requirements.txt
aws-cdk.core
aws-cdk.aws_elasticloadbalancing
aws-cdk.aws-ec2
aws-cdk.aws-autoscaling